32 Add import/upload markup
Added a the markdown to select a file and upload it. It still requires a
bit of clean up, but it is in a state where we can implement the actual
upload.

Note: The modal is mounted/unmounted when it's closed/opened.This was
done so that we don't have to deal with stale local state.Tearing
down/bringin up a component is expensive but the assumption is this
component will not be often seen.