32 add upload decks (#33)
* 32 Add import/upload markup

Added a the markdown to select a file and upload it. It still requires a
bit of clean up, but it is in a state where we can implement the actual
upload.

Note: The modal is mounted/unmounted when it's closed/opened.This was
done so that we don't have to deal with stale local state.Tearing
down/bringin up a component is expensive but the assumption is this
component will not be often seen.

* 32 Add upload state management

Added uploaded call to backend with accompanying state management.The UI
for the visual states will be done in an upcoming commit.

* 32 Add UI for states

Add the ui for the 3 states.Also added an action so that when the modal
is unmounted the ui states are re-initialized.This makes sure when it is
mounted again, a fresh state is displayed.

* 32 Moved the modal toggle button to modal component.

* 32 Make sure not all headers expect json content

* 32 Test for modal and update for decks page tests.

* 32 Remove inline styles

Replaced the inline styles with existing/new utility classes.The import
button required a css rule more specific than one used by semantic ui
and for that a new style sheet was created, components.css. The new
stylesheet will be used for styles specific to the re-usable components.

* 32 Remove console log