#21 Update types
The commit includes the following changes
1. Remove readonly prefix
2. Add __type__ field to entity typesS
3. Remove PostRequest type

Data mutability is not really a danger, since all updates are done in an
immutable manner. Removing the prefix saves us from needing to type it
for every new type definition.

Having the __type__ field gives a simple manner of discriminating if an
entity has already been saved/formed in db or is to be newly created.
Having the field, allows us to give all entities an id field which makes
live much easier when dealing with data in the components and reducer.

The PostRequest was simply not needed after making the changes above.