21 refactor newdeck and deck components (#41)
* #21 Update types

The commit includes the following changes
1. Remove readonly prefix
2. Add __type__ field to entity typesS
3. Remove PostRequest type

Data mutability is not really a danger, since all updates are done in an
immutable manner. Removing the prefix saves us from needing to type it
for every new type definition.

Having the __type__ field gives a simple manner of discriminating if an
entity has already been saved/formed in db or is to be newly created.
Having the field, allows us to give all entities an id field which makes
live much easier when dealing with data in the components and reducer.

The PostRequest was simply not needed after making the changes above.

* #21 Create a base type for each entity that is compatible with the initial and formed types.

* #21 Create a deck form component to be used in DeckPage and DeckNew components.

* #21 Implement DeckForm in DeckPage and DeckNew components.

* #21 Update the reset deck reducer so that the data fetching for the DeckPage is always performed when it is displayed.

* #21 Clean up fo CardForm.

* #21 Update tests

Added tests for the DeckForm component. The tests were written as unit
tests, so they are much cleaner and easier to follow. The NewDeck
component test are very minimal because most it those tests are now
under DeckForm.

* #21 Remove unused component