45 update layout (#58)
* #45 New layout structure

Added the structure of the new layout.It is lacking a lot of things,
but it is the foundation on which all changes will be built.

* #45 Finish structure of main sections

Ensured that the spacing of the main sections, i.e. sidebar, decks list,
decks header, and container of deck details are correct.

Added a border around each section, just to see clearly how they are
structured. All the inline styles will be removed.

* #45 Display the newly updated components.

* #45 Fix tests and use wrapper components.

* #45 Create section and section header components, along with a bit of updates to others.

* #45 Used new component to clean up the div soup.

* #46 Remove inlined styles and wrapper components.

* #45 Fix linter errors and deck details test.

* #45 Use wrapper component in deck test and remove margin from wrapper card.