64 Move dialog open state to parent
Move the dialog toggle state to its parent.Unlike the component that use
modal, the dialog does not operate independetly from its parent, i.e. it
is not mounted on a route so it is easier to control it in parent
component.

This commit also includes a little clean-up, the most important being
moving the border bottom css rule to sp card header component which
gives us a bit of visual consistency.