64 Display first deck by default
This change ensures that when decks are loaded the details of the first
decks are displayed by default.Without this, the DeckDetails component
would be empty until user clicks on a deck.

This commit also includes updating the loadStatus of each deck when the
list of decks is retrieved, this ensures that no unecessary calls to the
API are performed by the DeckDetails component and also removes the
jitter seen when navigating from one deck to another.