82 #82 Perform logout in Sidebar
Logging out had too many redirects so it was very jaggy.Perfomring the
logout in the sidebar removes all of that.

A user is logged out regardless of the result of the logut call.The
token is removed from localStorage so they would not be able to access
the app without logging in again.