28 Update login and signup forms
Update the login and signup forms so that they both display the form
validation errors when form is submitted, and the look of both is now
consitent with rest of app.

This commit is also includes changes to the signup form so that a user
can now provide their first name and last name and form also validates
their password via a confirm password input.