79 Display not found for unfound deck details
Display a not found error when the backend returns 404 for /decks/:id
calls. This is an edge case becuase the expectation is that a user
navigates to the deck details page but clicking on a deck card, i.e. an
existing card, but it is not an impossible state because one can try to
navigate to a card using url bar.