79 Remove unfound logic
The only way to reach details page of a deck is to click on it, i.e. it
has be present, so there is no need to handle the case where a user
tries to click on a non-existent deck, also since the decks are always
loaded when the decks sidebar is mounted we will always have the deck
details.Beforehand, the api did not return all the details necessary to
display the deck details component, but the is no longer the case.