79 add 404 page (#94)
* 79 Add not found component.

* 79 Display not found for unfound deck details

Display a not found error when the backend returns 404 for /decks/:id
calls. This is an edge case becuase the expectation is that a user
navigates to the deck details page but clicking on a deck card, i.e. an
existing card, but it is not an impossible state because one can try to
navigate to a card using url bar.

* 79 Remove unfound logic

The only way to reach details page of a deck is to click on it, i.e. it
has be present, so there is no need to handle the case where a user
tries to click on a non-existent deck, also since the decks are always
loaded when the decks sidebar is mounted we will always have the deck
details.Beforehand, the api did not return all the details necessary to
display the deck details component, but the is no longer the case.